Make Peformance/StartWith, EndWith, and StringReplacement much more robust
All of these cops now know how to handle all the various types of string
escapes in Ruby string literal syntax correctly. They work on a wider variety
of regex literals. Backslashes are handled correctly (which was not true in
some cases before).

EndWith incorrectly treated \Z as anchoring to the end of a string before;
rather, \z anchors to the end of a string. \Z anchors to the end of a string,
OR before the final newline if there is one. This means that converting a \Z
regex to `end_with?` could potentially break code.

Therefore, EndWith has been changed to look for \z rather than \Z.